requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-allauth==0.57.0
django-otp==1.2.0
qrcode==7.4.2
requests==2.31.0
sentry-sdk==1.38.0
django-sitemaps==1.0.0
django-robots==5.2
django-mptt==0.14.0
django-filter==23.5
django-countries==7.6.1
phonenumbers==8.13.26
django-phonenumber-field==7.1.0
django-cacheops==6.1
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
python-decouple==3.8
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
django-stubs==4.2.7